moved path check to utils
